add file info to context
